move User* into G4User sublibrary for clarity
