+++ Biasing/include/Biasing/EcalDarkBremFilter.h
@@ -23,7 +23,7 @@
 /*~~~~~~~~~~~~*/
 /*   SimCore  */
 /*~~~~~~~~~~~~*/
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 namespace biasing {
