+++ Biasing/include/Biasing/TaggerHitFilter.h
@@ -5,7 +5,7 @@
 #include <string>
 
 //~~ SimCore ~~//
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 //~~ Framework ~~//
 #include "Framework/Configure/Parameters.h"
