+++ Biasing/include/Biasing/TargetENProcessFilter.h
@@ -11,7 +11,7 @@
 /*~~~~~~~~~~~~~*/
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 /*~~~~~~~~~~*/
 /*   Core   */
