+++ Biasing/include/Biasing/Utility/PartialEnergySorter.h
@@ -9,7 +9,7 @@
 /*~~~~~~~~~~~~~*/
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 // Forward declarations
 class G4Step;
