+++ Biasing/include/Biasing/Utility/TrackProcessFilter.h
@@ -9,7 +9,7 @@
 /*~~~~~~~~~~~~~*/
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 namespace biasing {
 namespace utility {
