+++ Biasing/src/Biasing/DeepEcalProcessFilter.cxx
@@ -13,8 +13,8 @@
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
 #include "SimCore/G4User/VolumeChecks.h"
-#include "SimCore/UserEventInformation.h"
-#include "SimCore/UserTrackInformation.h"
+#include "SimCore/G4User/UserEventInformation.h"
+#include "SimCore/G4User/UserTrackInformation.h"
 
 namespace biasing {
