+++ Biasing/src/Biasing/NonFiducialFilter.cxx
@@ -17,8 +17,8 @@
 /*~~~~~~~~~~~~~*/
 #include "SimCore/G4User/PtrRetrieval.h"
 #include "SimCore/G4User/VolumeChecks.h"
-#include "SimCore/UserEventInformation.h"
-#include "SimCore/UserTrackInformation.h"
+#include "SimCore/G4User/UserEventInformation.h"
+#include "SimCore/G4User/UserTrackInformation.h"
 
 namespace biasing {
