+++ Biasing/src/Biasing/TargetDarkBremFilter.cxx
@@ -11,7 +11,7 @@
 
 #include "G4DarkBreM/G4APrime.h"           //checking if particles match A'
 #include "G4Electron.hh"                   //to check if track is electron
-#include "SimCore/UserTrackInformation.h"  //make sure A' is saved
+#include "SimCore/G4User/UserTrackInformation.h"  //make sure A' is saved
 
 namespace biasing {
