+++ SimCore/include/SimCore/G4User/EventAction.h
@@ -20,7 +20,7 @@
 /*~~~~~~~~~~~~~*/
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 // Forward declarations
 class G4Event;
