+++ SimCore/include/SimCore/G4User/StackingAction.h
@@ -20,7 +20,7 @@
 /*~~~~~~~~~~~~~*/
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 namespace simcore {
 namespace g4user {
