+++ SimCore/include/SimCore/G4User/SteppingAction.h
@@ -19,7 +19,7 @@
 /*~~~~~~~~~~~~~*/
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 namespace simcore {
 namespace g4user {
