+++ SimCore/include/SimCore/G4User/TrackingAction.h
@@ -22,7 +22,7 @@
 /*~~~~~~~~~~~~~*/
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 namespace simcore::g4user {
