+++ SimCore/include/SimCore/G4User/UserPrimaryParticleInformation.h
