+++ SimCore/src/SimCore/G4User/PrimaryGeneratorAction.cxx
@@ -16,8 +16,8 @@
 /*   SimCore   */
 /*~~~~~~~~~~~~~*/
 #include "SimCore/Generators/PrimaryGenerator.h"
-#include "SimCore/UserEventInformation.h"
-#include "SimCore/UserPrimaryParticleInformation.h"
+#include "SimCore/G4User/UserEventInformation.h"
+#include "SimCore/G4User/UserPrimaryParticleInformation.h"
 
 /*~~~~~~~~~~*/
 /*   ROOT   */
