+++ SimCore/src/SimCore/G4User/TrackingAction.cxx
@@ -2,9 +2,9 @@
 
 // LDMX
 #include "SimCore/G4User/TrackMap.h"
-#include "SimCore/UserPrimaryParticleInformation.h"
-#include "SimCore/UserRegionInformation.h"
-#include "SimCore/UserTrackInformation.h"
+#include "SimCore/G4User/UserPrimaryParticleInformation.h"
+#include "SimCore/G4User/UserRegionInformation.h"
+#include "SimCore/G4User/UserTrackInformation.h"
 
 // Geant4
 #include "G4PrimaryParticle.hh"
