+++ SimCore/src/SimCore/G4User/UserAction.cxx
@@ -1,5 +1,5 @@
 
-#include "SimCore/UserAction.h"
+#include "SimCore/G4User/UserAction.h"
 
 #include "SimCore/G4User/TrackingAction.h"
