+++ SimCore/src/SimCore/Generators/LHEPrimaryGenerator.cxx
@@ -17,7 +17,7 @@
 // LDMX
 #include "Framework/Configure/Parameters.h"
 #include "Framework/Exception/Exception.h"
-#include "SimCore/UserPrimaryParticleInformation.h"
+#include "SimCore/G4User/UserPrimaryParticleInformation.h"
 
 namespace simcore {
 namespace generators {
