+++ SimCore/src/SimCore/Geo/AuxInfoReader.cxx
@@ -6,7 +6,7 @@
 #include "Framework/Exception/Exception.h"
 #include "SimCore/MagneticFieldMap3D.h"
 #include "SimCore/MagneticFieldStore.h"
-#include "SimCore/UserRegionInformation.h"
+#include "SimCore/G4User/UserRegionInformation.h"
 #include "SimCore/VisAttributesStore.h"
 
 // Geant4
