+++ SimCore/src/SimCore/Simulator.cxx
@@ -25,7 +25,7 @@
 #include "SimCore/G4User/TrackingAction.h"
 #include "SimCore/Generators/PrimaryGenerator.h"
 #include "SimCore/SDs/SensitiveDetector.h"
-#include "SimCore/UserEventInformation.h"
+#include "SimCore/G4User/UserEventInformation.h"
 
 /*~~~~~~~~~~~~~~*/
 /*    Geant4    */
